Move retention code into convert processor file.
